Allow example code to be read from file
This is actually a pretty major change because of how hardcoded the
examples are. It completely rewrites the parser as hand-coded RD
instead of PEG, adds strings to the language, adds a `print_num`
function to the language, adds a concept of a module and executes
a `main` function automatically. In general I think it is much more
of a proper language now (but still a toy). Before it could really
only execute the provided examples and would be hard to extend
without doing all this work first.